+++ scripts/move_to_height.py
@@ -1,13 +1,11 @@
 import time
 import RPi.GPIO as GPIO
+from constants import UP_PIN, DOWN_PIN
 
 # Calibration data
 LOWEST_HEIGHT = 23.7  # inches
 HIGHEST_HEIGHT = 54.5  # inches
 UP_RATE = 0.54  # inches per second (from calibration)
-
-UP_PIN = 17   # BCM numbering, physical pin 11
-DOWN_PIN = 27 # BCM numbering, physical pin 13
 
 GPIO.setmode(GPIO.BCM)